Fixed support for multiple coverage files
In case of multiple coverage files, the plugin tries to copy them to /jacoco1 where it usually does not have any access. Generating the names of the additional files (with number appended) in the same way as the original file name fixed it for me. Then I at least get a working summary of all coverages.

Keep up the good work with the plugin, we really need a new, flexible coverage plugin for Jenkins.