Show diff in coverage for change request builds
It is convinient to see the code coverage diff right in Jenkins. If our
build is going for a multibranch change request job  - we take the
target branch commit that is used for our build. Then we search for the
last fixed amount of builds the build that was built for this commit. If
we find it - we check if it has code coverage information. If it has
lines coverage - we calculate diff between the current code coverage and
this target branch build coverage.