Fix #150. Remove wars about source files with the same names
The Jacoco fix in 4081720909b brought new code that tries to find the
source file in case other ways didn't work. To do this it creates a
map of file names to look for. And it logs warnings in case of any
duplicates. Becase there may be a lot of them a log would quite apace,
we remove the warnings.