CoverageNode: Fix mergePath for node created from cobertura report
When a coverage node has been created from a cobertura report, the mergePath
method was returning a wrong path if the provided local path contains any
subpaths in it. Those subpaths were prepended to the local path thus doubling
them in the merged path (for instance "a/b/c" was transformed to "a/b/a/b/c").

As a consequence, FileCoverageNode.getPath was returning a wrong file path which
prevented any source file painting and display in the Jenkins UI.

Fix #456